Move apiCall.tsx to a general Utils folder in frontend, so other devs can also use it
