+++ frontend/src/components/Authentication/SetUp2FA.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import SingleHeader from './Pages/SingleHeader.tsx';
-import { getQRCode } from './apiCalls.tsx';
+import { getQRCode } from '../Utils/apiCalls.tsx';
 import Enable2FA from './Enable2FA.tsx';
 
 const SetUp2FA = () => {
